% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geth.R
\name{citation_speed}
\alias{citation_speed}
\title{Estimate citation speed for publications}
\usage{
citation_speed(data, n = 10)
}
\arguments{
\item{data}{A tidy data frame created by \code{clean_scopus()}.}

\item{n}{Number of rows to return (default = 10).}
}
\value{
A tidy data frame containing publication year,
journal name, and estimated months per citation.
}
\description{
Estimates the average number of months required for a publication
to receive one citation.
}
\author{
Jose V. Die
}
