% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geth.R
\name{clean}
\alias{clean}
\title{Read & Clean data from a File Dowloaded from the Scopus Database}
\usage{
clean(fileName)
}
\arguments{
\item{file}{a file downloaded from Scopus}
}
\description{
This function reads and performs tidy data on a file dowloaded
from the Scopus Database. For more info on how to dowload the file, please
read the tutorial of this package in https://github.com/jdieramon/hIndex
}
\author{
Jose V. Die
}
