% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geth.R
\name{clean_scopus}
\alias{clean_scopus}
\title{Read and clean data downloaded from Scopus}
\usage{
clean_scopus(fileName)
}
\arguments{
\item{fileName}{Character string indicating the path to a CSV file
downloaded from Scopus.}
}
\value{
A tidy data frame containing publication metadata and
yearly citation counts.
}
\description{
Reads a CSV file exported from the Scopus database and performs
basic cleaning and formatting operations.
}
\details{
The function reformats column names and removes unnecessary rows
and columns generated during the Scopus export process.
}
\author{
Jose V. Die
}
