% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geth.R
\name{df_h}
\alias{df_h}
\title{Simple dataset : year and h index}
\usage{
df_h(file, a, b, h)
}
\arguments{
\item{file}{a tidy data frame with papers in rows and other info in colums;}

\item{a}{starting year.}

\item{b}{end year. It is usually the current year.}

\item{h}{the h-index that corresponds to the starting year (by default=0).}
}
\description{
Core function used by \code{\link{h.model}}.
}
\author{
Jose V. Die
}
