% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geth.R
\name{expected_citations}
\alias{expected_citations}
\title{Estimate the number of expected citations in a period of time}
\usage{
expected_citations(dat, 6)
}
\arguments{
\item{file}{a tidy data frame}

\item{term}{starting year.}
}
\description{
This function estimated the number of expected citations in a period of time 
based on the time that takes each paper to get 1 citation.
}
\author{
Jose V. Die
}
