% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geth.R
\name{format1cite}
\alias{format1cite}
\title{Format the file and estimate the time that takes to get 1 citation}
\usage{
format1cite(file)
}
\arguments{
\item{file}{a tidy data frame}
}
\value{
a tidy data frame that will be used by \code{plot1cite}
}
\description{
Format the file and estimate the time that takes to get 1 citation
}
\author{
Jose V. Die
}

