% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geth.R
\name{get1cite}
\alias{get1cite}
\title{Format the file and estimate the time (in months) that takes to get 1 citation}
\usage{
get1cite(file, n = 10)
}
\arguments{
\item{file}{a tidy data frame}
}
\value{
a tidy data frame that will be used by \code{plot1cite}
}
\description{
Format the file and estimate the time (in months) that takes to get 1 citation
}
\author{
Jose V. Die
}
