% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geth.R
\name{h.model}
\alias{h.model}
\title{Model the h-index over time}
\usage{
h.model(file, a, b, h)
}
\arguments{
\item{file}{a tidy data frame with papers in rows and other info in columns;}

\item{a}{starting year.}

\item{b}{end year. It is usually the current year.}

\item{h}{the h-index that corresponds to the starting year (by default=0).}
}
\description{
This function makes a linear regression model of the h-index value for a given 
author over the years.
}
\author{
Jose V. Die
}
