% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geth.R
\name{plot1cite}
\alias{plot1cite}
\title{Plot the time that takes to get 1 citation for a set of publications}
\usage{
plot1cite(file)
}
\arguments{
\item{file}{a tidy data frame produced by \code{format1cite}}
}
\description{
This function takes the file produced by the \code{format1cite} and plots
it up.
}
\author{
Jose V. Die
}
