% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geth.R
\name{plot_h}
\alias{plot_h}
\title{Plot h-index trajectory over time}
\usage{
plot_h(data, start_year = NULL, end_year = NULL, h = 0)
}
\arguments{
\item{data}{A tidy data frame created by \code{clean_scopus()}.}

\item{start_year}{Optional starting year for the analysis.}

\item{end_year}{Optional ending year for the analysis.}

\item{h}{Initial h-index value (default = 0).}
}
\value{
A ggplot object representing the h-index trajectory.
}
\description{
Visualizes the evolution of the h-index across years.
}
\author{
Jose V. Die
}
