% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geth.R
\name{predict_citations}
\alias{predict_citations}
\title{Estimate future citations over a time period}
\usage{
predict_citations(data, term = 12)
}
\arguments{
\item{data}{A tidy data frame created by \code{clean_scopus()}.}

\item{term}{Numeric value indicating the prediction period in months.}
}
\value{
A tidy data frame containing estimated future citations.
}
\description{
Estimates the expected number of citations over a specified
time window based on historical citation speed.
}
\details{
Citation projections are based on a simple linear approximation:

\deqn{
expected\_citations =
\left\lfloor
\frac{term}{months\_per\_citation}
\right\rfloor
}
}
\seealso{
\code{citation_speed}
}
\author{
Jose V. Die
}
