% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geth.R
\name{geth}
\alias{geth}
\title{Get the h-index From the Scopus Database}
\usage{
geth(file, h = 0)
}
\arguments{
\item{file}{a tidy data frame with papers in row and other info in colums
where column 7 corresponds to the starting year.}

\item{h}{an integer number with the h-index for the starting year.}
}
\description{
This function assumes tidy data from Scopus (file) and generates the h-index
over years (from the first year [h=0] until the most recent year [h=x]).
}
\details{
This function calculates cumulative number of citations starting with the
first year of your publication (which is assumed in the function to be h=0)
}
\author{
Jose V. Die
}
\seealso{
\code{rowSums}
}

