% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geth.R
\name{h.plot}
\alias{h.plot}
\title{Plot the h-trajectory}
\usage{
h.plot(file)
}
\arguments{
\item{file}{a tidy data frame with papers in rows and other info in colums;
column 7 corresponds to the first year of h=0.}
}
\description{
This function takes the h-index values for a given author over the years produced
by the \code{geth} function and plots them up.
}
\author{
Jose V. Die
}

