% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geth.R
\name{h.plot}
\alias{h.plot}
\title{Plot the h-trajectory}
\usage{
h.plot(file, a, b)
}
\arguments{
\item{file}{a tidy data frame with papers in rows and other info in colums;}

\item{a}{starting year, so h=0 for that year.}

\item{b}{end year. It is usually the current year.}
}
\description{
This function takes the h-index values for a given author over the years produced
by the \code{geth} function and plots them up.
}
\author{
Jose V. Die
}

